Bug 1614846 follow-up - Use size_t instead of uint32_t to un-break Windows 64-bit builds. rs=bustage-fix

+++ mailnews/imap/src/nsImapOfflineSync.cpp
@@ -123,7 +123,7 @@ bool nsImapOfflineSync::AdvanceToNextServer() {
     rv = accountManager->GetAllServers(m_allServers);
     NS_ENSURE_SUCCESS(rv, false);
   }
-  uint32_t serverIndex = 0;
+  size_t serverIndex = 0;
   if (m_currentServer) {
     serverIndex = m_allServers.IndexOf(m_currentServer);
     if (serverIndex == m_allServers.NoIndex) {
